requirements.txt
openai
requests 
pillow 
numpy